Refactor : Modify folder name to entity in review
