+++ src/review/repository/review.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Review } from '../entities/review.entity';
+import { Review } from '../entity/review.entity';
 
 @Injectable()
 export class ReviewRepository {
