+++ src/review/review.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { ReviewService } from './review.service';
 import { ReviewController } from './review.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Review } from './entities/review.entity';
-import { RestaurantReviewAggregation } from './entities/restaurant-review-aggregation.entity';
+import { Review } from './entity/review.entity';
+import { RestaurantReviewAggregation } from './entity/restaurant-review-aggregation.entity';
 import { ReviewRepository } from './repository/review.repository';
 import { RestaurantReviewAggregationRepository } from './repository/restaurant-review-aggregation.repository';
 import { RestaurantModule } from 'src/restaurant/restaurant.module';
